edited keyserver

+++ keyserver.py
@@ -29,20 +29,28 @@ while True:
                 conn = psycopg2.connect(host="localhost", port=DATABASE_PORT, dbname="fastchatdb", user="postgres", password="AshwinPostgre")
                 cur = conn.cursor()
                 command = data["command"]
+
                 if command == "STORE":
                     username = data["username"]
                     key = data["key"]
                     type = data["type"]
-                    # cur.execute(f"SELECT * FROM KEYSERVER WHERE username='{username}' AND type='{type}'")
-                    # if(len(cur.fetchall()) > 0):
-                    #     print("[DEBUG]: Error: User already exists")
-                    #     sockfd.send(to_send({"command": "ERROR", "msg": "User already exists!\n"}))
-                    # else:
-                    cur.execute(f"DELETE FROM KEYSERVER WHERE username='{username}'")
-                    cur.execute(f"INSERT INTO KEYSERVER VALUES ('{username}', '{key}', '{type}')")
-                    conn.commit()
-                    print("[DEBUG]: Successfully stored")
-                    sockfd.send(to_send({"command": "INFO", "msg": "Successfully stored!\n"}))
+                    if(type=="group"):
+                        cur.execute(f"DELETE FROM KEYSERVER WHERE username='{username}'")
+                        cur.execute(f"INSERT INTO KEYSERVER VALUES ('{username}', '{key}', '{type}')")
+                        conn.commit()
+                        print("[DEBUG]: Successfully stored")
+                        sockfd.send(to_send({"command": "INFO", "msg": "Successfully stored!\n"}))            
+                    else:
+                        cur.execute(f"SELECT * FROM KEYSERVER WHERE username='{username}' AND type='{type}'")
+                        if(len(cur.fetchall()) > 0):
+                            print("[DEBUG]: Error: User already exists")
+                            sockfd.send(to_send({"command": "ERROR", "msg": "User already exists!\n"}))
+                        else:
+                            cur.execute(f"INSERT INTO KEYSERVER VALUES ('{username}', '{key}', '{type}')")
+                            conn.commit()
+                            print("[DEBUG]: Successfully stored")
+                            sockfd.send(to_send({"command": "INFO", "msg": "Successfully stored!\n"}))
+                    
                 elif command == "RETRIEVE":
                     username = data["username"]
                     type = data["type"]
